Add router middleware
